Use a better file name date stamp format
Thank you for adding support for long file names and sane and readable dates.

However, your new date format, I consider it to be a regression.

Your old date format was:
- unambiguous (it was a specialized encoding of UTC, easily matched with a regex)
- unreadable (it was an encoding)

Your new date format is:
- ambiguous (multiple other sensor platforms use the same format, it is not distinguishable from them, thus can now only be inferred as an ambiguous date time with respect to UTC)
- readable

My proposed date format:
- is unambiguous (the trailing 'Z' character is the standard notation for UTC/Zulu time)
- is readable (although for a beginner, slightly less so given the replacement of `_` with `T`
- follows the ISO8601 standard: https://en.wikipedia.org/wiki/ISO_8601


You make one of the few sensors on the market that has some notion of the real-time a recording occurred. Please don't obscure this information. I'm confident there's a balance between readability for novices and interpretability for programs. 

I'm aware you also encode unique information in the headers. In my opinion, if I have to read bytes in a file header to understand a datestamp then the problem is already too hard - especially for a novice who is parsing a date in R.